requirements.txt
psycopg2-binary
mysqlclient
sqlalchemy-utils
sqlalchemy
